# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Elrond Portal API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://elrond-portal.test"
        }
    ],
    "paths": {
...
        "/api/v1/kingdom": {
            "get": {
                "tags": [
                    "Kingdom"
                ],
                "summary": "Retorna informações dos reinos",
                "operationId": "0544b6a2f78182d61a46aec212e4d8bc",
                "responses": {
                    "200": {
                        "description": "Informações dos reinos retornadas com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "kingdoms": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Conselho Exemplo"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Descrição do Conselho Exemplo"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Kingdom"
                ],
                "summary": "Cria um novo reino",
                "operationId": "2ad06522b8b45ea59326fad225ffc158",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Reino Exemplo"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descrição do Reino Exemplo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reino criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "kingdom": {
                                            "type": "object",
                                            "example": {
                                                "id": 1,
                                                "name": "Reino Exemplo",
                                                "description": "Descrição do Reino Exemplo"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                }
            }
        },
        "/api/v1/role": {
            "get": {
                "tags": [
                    "Role"
                ],
                "summary": "Retorna informações dos papéis",
                "operationId": "4cebb0ba7da6403f5178d7221945b6cb",
                "responses": {
                    "200": {
                        "description": "Informações dos papéis retornadas com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "roles": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Papel Exemplo"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Descrição do Papel Exemplo"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Role"
                ],
                "summary": "Cria um novo papel",
                "operationId": "ebf680c8b7a8d46aeefccc7548183022",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Papel Exemplo"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descrição do Papel Exemplo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Papel criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "role": {
                                            "type": "object",
                                            "example": {
                                                "id": 1,
                                                "name": "Papel Exemplo",
                                                "description": "Descrição do Papel Exemplo"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}